Smart Eyewear Reminder Simulation: Arduino Implementation
his Arduino code simulates the functionality of a smart eyewear reminder device. 
It uses LEDs to represent the presence of glasses and a push button to simulate going out of the house. 
When the button is pressed, the code checks for the presence of glasses. If glasses are not present, 
it activates another LED to simulate sending a notification.