ESP32 Base Station (Home)
WiFi Setup (setup_wifi): This function connects the ESP32 to the local WiFi network using the provided SSID and password.

MQTT Callback (callback): This function is called whenever a message is received on the subscribed MQTT topic. If the message indicates whether the glasses are detected at home or not, it prints the appropriate message.

MQTT Reconnect (reconnect): This function attempts to reconnect to the MQTT broker if the connection is lost.

Setup Function (setup): In the setup function, we initialize serial communication, set up WiFi, and set up MQTT client. We also set up the callback function and subscribe to the MQTT topic where the glasses' status will be published.

Loop Function (loop): In the loop function, we check if the MQTT client is connected. If not, we attempt to reconnect. We also continuously listen for incoming MQTT messages.